Refactor chat-message-tool
