+++ frontend/src/components/chat/scroll-to-bottom.tsx
@@ -1,0 +1,11 @@
+import { useEffect, useRef } from "react";
+
+export function ScrollToBottom() {
+  const messagesEndRef = useRef<HTMLDivElement>(null);
+
+  useEffect(() => {
+    messagesEndRef.current?.scrollIntoView({ behavior: "smooth" });
+  }, []);
+
+  return <div ref={messagesEndRef} />;
+}
